# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para la plataforma LogiRed de transporte de paquetes",
        "title": "LogiRed API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/proposals/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proposals"
                ],
                "summary": "Obtener una propuesta por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la propuesta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "datos de la propuesta",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "propuesta no encontrada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Solo el conductor propietario puede eliminar su propuesta",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proposals"
                ],
                "summary": "Eliminar una propuesta",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la propuesta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "mensaje de éxito",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "no autenticado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "propuesta no encontrada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/rides/city/{city}": {
            "get": {
                "description": "Devuelve viajes activos (estado 6) en la ciudad especificada",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rides"
                ],
                "summary": "Obtener viajes disponibles por ciudad",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nombre de la ciudad",
                        "name": "city",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "lista de viajes",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ciudad no especificada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/rides/driver/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Devuelve todos los viajes que el conductor ha aceptado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rides"
                ],
                "summary": "Obtener viajes asignados al conductor autenticado",
                "responses": {
                    "200": {
                        "description": "lista de viajes",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "no autenticado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/rides/{idride}/proposals": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proposals"
                ],
                "summary": "Obtener todas las propuestas de un viaje",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del viaje",
                        "name": "idride",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "lista de propuestas",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rides/{id}": {
            "get": {
                "description": "Devuelve los detalles de un viaje específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rides"
                ],
                "summary": "Obtener un viaje por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del viaje",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "datos del viaje",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "viaje no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Escribe \"Bearer \u003ctoken\u003e\" para autenticarte",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}